pyproject.toml
[tool.poetry]
name = "usersnack"
version = "0.1.0"
description = "A FastAPI backend for a pizza ordering service."
authors = ["Kilo Code"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
psycopg = {extras = ["binary"], version = "^3.1.18"}
alembic = "^1.13.0"
structlog = "^24.1.0"
slowapi = "^0.1.9"
dependency-injector = "^4.41.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"
testcontainers-postgres = "^0.0.1a5"
ruff = "^0.2.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."